Move webview context menu
