+++ src/actions.js
@@ -68,6 +68,8 @@ export const UPDATES_READY = 'updates/ready';
 export const UPDATES_UPDATE_DOWNLOADED = 'updates/update-downloaded';
 export const WEBVIEW_CERTIFICATE_DENIED = 'webview/certificate-denied';
 export const WEBVIEW_CERTIFICATE_TRUSTED = 'webview/certificate-trusted';
+export const WEBVIEW_CONTEXT_MENU_POPPED_UP = 'webview/context-menu-popped-up';
+export const WEBVIEW_MISSPELLED_WORDS_REQUESTED = 'webview/misspelled-words-requested';
 export const WEBVIEW_DID_NAVIGATE = 'webview/did-navigate';
 export const WEBVIEW_EDIT_FLAGS_CHANGED = 'webview/edit-flags-changed';
 export const WEBVIEW_FAVICON_CHANGED = 'webview/favicon-changed';
