--- src/hooks/useCorrectionsForMisspelling.js
@@ -1,4 +1,0 @@
-import { useCallableSaga } from '../components/SagaMiddlewareProvider';
-import { getCorrectionsForMisspelling } from '../sagas/spellChecking';
-
-export const useCorrectionsForMisspelling = () => useCallableSaga(getCorrectionsForMisspelling, []);
