+++ src/main/sagas/rootWindow.js
@@ -1,6 +1,6 @@
 import path from 'path';
 
-import { app, BrowserWindow, Menu, dialog } from 'electron';
+import { app, BrowserWindow, Menu, dialog, ipcMain, shell, clipboard } from 'electron';
 import { t } from 'i18next';
 import { takeEvery, select, put, call, getContext, spawn } from 'redux-saga/effects';
 import { createSelector } from 'reselect';
@@ -40,6 +40,9 @@ import {
 	WEBVIEW_CERTIFICATE_DENIED,
 	WEBVIEW_CERTIFICATE_TRUSTED,
 	WEBVIEW_FOCUS_REQUESTED,
+	WEBVIEW_CONTEXT_MENU_POPPED_UP,
+	WEBVIEW_SPELL_CHECKING_DICTIONARY_FILES_CHOSEN,
+	WEBVIEW_SPELL_CHECKING_DICTIONARY_TOGGLED,
 } from '../../actions';
 import { eventEmitterChannel, storeChangeChannel } from '../channels';
 import { getTrayIconPath, getAppIconPath } from '../icons';
@@ -50,6 +53,9 @@ import {
 	selectIsTrayIconEnabled,
 	selectIsShowWindowOnUnreadChangedEnabled,
 	selectServers,
+	selectSpellCheckingDictionaries,
+	selectInstalledSpellCheckingDictionariesDirectoryPath,
+	selectFocusedWebContents,
 } from '../selectors';
 import { validateServerUrl, ValidationResult } from '../servers';
 
@@ -360,6 +366,192 @@ function *watchRootWindow(rootWindow, store) {
 
 			yield put({ type: DEEP_LINKS_SERVER_ADDED, payload: url });
 		}
+	});
+
+	yield takeEvery(WEBVIEW_CONTEXT_MENU_POPPED_UP, function *({ payload: params }) {
+		const dictionaries = yield select(selectSpellCheckingDictionaries);
+		const dictionariesDirectoryPath = yield select(selectInstalledSpellCheckingDictionariesDirectoryPath);
+
+		const getCorrectionsForMisspelling = (text) => new Promise((resolve) => {
+			ipcMain.once('get-corrections-for-misspelling-response', (event, response) => {
+				resolve(response);
+			});
+			rootWindow.webContents.send('get-corrections-for-misspelling', text);
+		});
+		const webContents = yield select(selectFocusedWebContents);
+
+		const createSpellCheckingMenuTemplate = ({
+			isEditable,
+			corrections,
+			dictionaries,
+		}) => {
+			if (!isEditable) {
+				return [];
+			}
+
+			const handleBrowserForLanguage = async () => {
+				const { filePaths } = await dialog.showOpenDialog(rootWindow, {
+					title: t('dialog.loadDictionary.title'),
+					defaultPath: dictionariesDirectoryPath,
+					filters: [
+						{ name: t('dialog.loadDictionary.dictionaries'), extensions: ['dic', 'aff'] },
+						{ name: t('dialog.loadDictionary.allFiles'), extensions: ['*'] },
+					],
+					properties: ['openFile', 'multiSelections'],
+				});
+
+				store.dispatch({ type: WEBVIEW_SPELL_CHECKING_DICTIONARY_FILES_CHOSEN, payload: filePaths });
+			};
+
+			return [
+				...corrections ? [
+					...corrections.length === 0
+						? [
+							{
+								label: t('contextMenu.noSpellingSuggestions'),
+								enabled: false,
+							},
+						]
+						: corrections.slice(0, 6).map((correction) => ({
+							label: correction,
+							click: () => webContents.replaceMisspelling(correction),
+						})),
+					...corrections.length > 6 ? [
+						{
+							label: t('contextMenu.moreSpellingSuggestions'),
+							submenu: corrections.slice(6).map((correction) => ({
+								label: correction,
+								click: () => webContents.replaceMisspelling(correction),
+							})),
+						},
+					] : [],
+					{ type: 'separator' },
+				] : [],
+				{
+					label: t('contextMenu.spellingLanguages'),
+					enabled: dictionaries.length > 0,
+					submenu: [
+						...dictionaries.map(({ name, enabled }) => ({
+							label: name,
+							type: 'checkbox',
+							checked: enabled,
+							click: ({ checked }) => store.dispatch({
+								type: WEBVIEW_SPELL_CHECKING_DICTIONARY_TOGGLED,
+								payload: { name, enabled: checked },
+							}),
+						})),
+						{ type: 'separator' },
+						{
+							label: t('contextMenu.browseForLanguage'),
+							click: handleBrowserForLanguage,
+						},
+					],
+				},
+				{ type: 'separator' },
+			];
+		};
+
+		const createImageMenuTemplate = ({
+			mediaType,
+			srcURL,
+		}) => (
+			mediaType === 'image' ? [
+				{
+					label: t('contextMenu.saveImageAs'),
+					click: () => webContents.downloadURL(srcURL),
+				},
+				{ type: 'separator' },
+			] : []
+		);
+
+		const createLinkMenuTemplate = ({
+			linkURL,
+			linkText,
+		}) => (
+			linkURL
+				? [
+					{
+						label: t('contextMenu.openLink'),
+						click: () => shell.openExternal(linkURL),
+					},
+					{
+						label: t('contextMenu.copyLinkText'),
+						click: () => clipboard.write({ text: linkText, bookmark: linkText }),
+						enabled: !!linkText,
+					},
+					{
+						label: t('contextMenu.copyLinkAddress'),
+						click: () => clipboard.write({ text: linkURL, bookmark: linkText }),
+					},
+					{ type: 'separator' },
+				]
+				: []
+		);
+
+		const createDefaultMenuTemplate = ({
+			editFlags: {
+				canUndo = false,
+				canRedo = false,
+				canCut = false,
+				canCopy = false,
+				canPaste = false,
+				canSelectAll = false,
+			} = {},
+		} = {}) => [
+			{
+				label: t('contextMenu.undo'),
+				role: 'undo',
+				accelerator: 'CommandOrControl+Z',
+				enabled: canUndo,
+			},
+			{
+				label: t('contextMenu.redo'),
+				role: 'redo',
+				accelerator: process.platform === 'win32' ? 'Control+Y' : 'CommandOrControl+Shift+Z',
+				enabled: canRedo,
+			},
+			{ type: 'separator' },
+			{
+				label: t('contextMenu.cut'),
+				role: 'cut',
+				accelerator: 'CommandOrControl+X',
+				enabled: canCut,
+			},
+			{
+				label: t('contextMenu.copy'),
+				role: 'copy',
+				accelerator: 'CommandOrControl+C',
+				enabled: canCopy,
+			},
+			{
+				label: t('contextMenu.paste'),
+				role: 'paste',
+				accelerator: 'CommandOrControl+V',
+				enabled: canPaste,
+			},
+			{
+				label: t('contextMenu.selectAll'),
+				role: 'selectall',
+				accelerator: 'CommandOrControl+A',
+				enabled: canSelectAll,
+			},
+		];
+
+		const props = {
+			...params,
+			corrections: yield call(getCorrectionsForMisspelling, params.selectionText),
+			dictionaries,
+		};
+
+		const template = [
+			...createSpellCheckingMenuTemplate(props),
+			...createImageMenuTemplate(props),
+			...createLinkMenuTemplate(props),
+			...createDefaultMenuTemplate(props),
+		];
+
+		const menu = Menu.buildFromTemplate(template);
+		menu.popup({ window: rootWindow });
 	});
 }
